Se han colocado las imagenes en una carpeta llamada: default para el juego tradicional. Se están mostrando las piezas "sprites" desde la clase GLTablero

+++ src/Casilla.h
@@ -18,10 +18,10 @@ public:
 		if (sum % 2) glColor3ub(rc, gc, bc); //para casillas pares un color mas claro//glColor3ub(23, 31, 161);
 		else glColor3ub(ro, go, bo);//glColor3ub(126, 128, 156);
 		glBegin(GL_POLYGON);
-		glVertex2f(i + 1, j + 1); //antihorario 
-		glVertex2f(i + 1, j);
-		glVertex2f(i, j);
-		glVertex2f(i, j + 1);
+		glVertex2f((float)i + 1, (float)j + 1); //antihorario 
+		glVertex2f((float)i + 1, (float)j);
+		glVertex2f((float)i, (float)j);
+		glVertex2f((float)i, (float)j + 1);
 		glEnd();
 		glEnable(GL_LIGHTING);
 	}
@@ -30,10 +30,10 @@ public:
 		glLineWidth(2);
 		glColor3ub(231, 0, 0);
 		glBegin(GL_LINES);
-		glVertex3f(i + 1, j + 1, 0.01f); glVertex3f(i + 1, j, 0.01f);
-		glVertex3f(i + 1, j, 0.01f);	 glVertex3f(i, j, 0.01f);
-		glVertex3f(i, j, 0.01f);		 glVertex3f(i, j + 1, 0.01f);
-		glVertex3f(i, j + 1, 0.01f);  	 glVertex3f(i + 1, j + 1, 0.01f);
+		glVertex3f((float)i + 1.0f, (float)j + 1, 0.01f); glVertex3f(i + 1.0f, j, 0.01f);
+		glVertex3f((float)i + 1.0f, (float)j, 0.01f);	 glVertex3f(i, j, 0.01f);
+		glVertex3f((float)i, (float)j, 0.01f);		 glVertex3f(i, j + 1.0f, 0.01f);
+		glVertex3f((float)i, (float)j + 1, 0.01f);  	 glVertex3f(i + 1.0f, j + 1.0f, 0.01f);
 		glEnd();
 	}
 };
